requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
mcp>=0.6.0
langchain-mcp-adapters>=0.1.0
chromadb>=0.4.0
dashscope>=1.14.0
